move business logic of watch command to another package
