+++ pkg/cmd/doc.go
@@ -1,0 +1,4 @@
+// The pkg/cmd package is where the actual business logic for warpforge commands resides.
+// This differs from the cmd/warpforge directory which is specifically a CLI application.
+// The purpose is to create a clear split between CLI concerns and business logic of warpforge.
+package cmd
